add DwordOverloader class
- enum class : DWORD 객체를 "DwordOverloader" 템플릿 클래스를 상속받는 class로 전환했다.
- Common.h에 불필요한 추가 코드 작성 없이 DWORD 연산이 가능하도록 한다.

+++ Client/X-MACHINA/include/GameFramework.h
@@ -39,7 +39,7 @@ public:
 
 
 public:
-	bool Init(HINSTANCE hInstance, LONG width, LONG height);
+	bool Init(HINSTANCE hInstance, short width, short height);
 	LRESULT ProcessMessage(HWND hWnd, UINT message, WPARAM wParam, LPARAM lParam);
 	void ProcessKeyboardMsg(HWND hWnd, UINT message, WPARAM wParam, LPARAM lParam);
 	void ProcessMouseMsg(HWND hWnd, UINT message, WPARAM wParam, LPARAM lParam);
